refactor(app): move internal auth to react
