--- app/portainer/settings/authentication/internal-auth/index.ts
@@ -1,1 +1,0 @@
-export { InternalAuthAngular, InternalAuth } from './InternalAuth';
