+++ app/react/portainer/settings/AuthenticationView/InternalAuth/Badge.tsx
