+++ app/react/portainer/settings/AuthenticationView/InternalAuth/index.ts
@@ -1,0 +1,1 @@
+export { InternalAuth } from './InternalAuth';
